requirements.txt
streamlit>=1.24
tensorflow>=2.10
pillow
numpy
